jan 25 2019 @ 6:50pm    looks like the segfaul t was fixed, for some reason it was in my BST.hpp    file in the deleteAll function out of all places...

+++ src/BST.hpp
@@ -257,9 +257,12 @@ private:
     // TODO: DONE
     static void deleteAll(BSTNode<Data> *n) {
 	BSTNode<Data>* current = n;
+	if(n == nullptr)
+		return;
 	if(current->left == nullptr){
-		if(current->right == nullptr)
+		if(current->right == nullptr){
 			delete current;
+		}
 		else
 			deleteAll(current->right);
 	}
